requirements.txt
crispy-bootstrap5==0.6
Django==4.0.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
django-taggit
django-ckeditor
Pillow
django-cors-headers